1: refactoring ok - 2: edit backend to set empty string if no image to a post
